Move RabbitMQ registration to dedicated function
